+++ PayPalWebPayments/src/test/java/com/paypal/android/paypalwebpayments/PayPalWebLauncherUnitTest.kt
@@ -4,7 +4,6 @@ import android.content.Intent
 import android.net.Uri
 import androidx.fragment.app.FragmentActivity
 import com.braintreepayments.api.BrowserSwitchClient
-import com.braintreepayments.api.BrowserSwitchException
 import com.braintreepayments.api.BrowserSwitchFinalResult
 import com.braintreepayments.api.BrowserSwitchOptions
 import com.braintreepayments.api.BrowserSwitchStartResult
@@ -16,8 +15,6 @@ import io.mockk.mockk
 import io.mockk.slot
 import junit.framework.TestCase.assertEquals
 import org.json.JSONObject
-import org.junit.Assert.assertSame
-import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -238,40 +235,6 @@ class PayPalWebLauncherUnitTest {
     }
 
     @Test
-    fun `completeAuthRequest() returns unknown error when browser switch fails`() {
-        val browserSwitchError = BrowserSwitchException("browser switch error")
-        val finalResult = mockk<BrowserSwitchFinalResult.Failure>(relaxed = true)
-        every { finalResult.error } returns browserSwitchError
-        every {
-            browserSwitchClient.completeRequest(intent, "pending request")
-        } returns finalResult
-
-        sut = PayPalWebLauncher("custom_url_scheme", liveConfig, browserSwitchClient)
-
-        val status = sut.completeAuthRequest(intent, "pending request")
-                as PayPalWebStatus.UnknownError
-        assertSame(browserSwitchError, status.error)
-    }
-
-    @Test
-    fun `completeAuthRequest() returns no result when request code is not for PayPal`() {
-        val browserSwitchResult = createCheckoutSuccessBrowserSwitchResult(
-            requestCode = BrowserSwitchRequestCodes.CARD_APPROVE_ORDER,
-            orderId = "fake-order-id",
-            payerId = "fake-payer-id"
-        )
-
-        every {
-            browserSwitchClient.completeRequest(intent, "pending request")
-        } returns browserSwitchResult
-
-        sut = PayPalWebLauncher("custom_url_scheme", liveConfig, browserSwitchClient)
-
-        val status = sut.completeAuthRequest(intent, "pending request")
-        assertTrue(status is PayPalWebStatus.NoResult)
-    }
-
-    @Test
     fun `completeCheckoutAuthRequest() parses successful checkout result`() {
         val browserSwitchResult = createCheckoutSuccessBrowserSwitchResult(
             requestCode = BrowserSwitchRequestCodes.PAYPAL_CHECKOUT,
@@ -349,7 +312,7 @@ class PayPalWebLauncherUnitTest {
     }
 
     @Test
-    fun `completeRequest() parses successful vault result`() {
+    fun `completeVaultAuthRequest() parses successful vault result`() {
         val browserSwitchResult = createVaultSuccessBrowserSwitchResult(
             requestCode = BrowserSwitchRequestCodes.PAYPAL_VAULT,
             setupTokenId = "fake-setup-token-id",
@@ -360,13 +323,13 @@ class PayPalWebLauncherUnitTest {
         } returns browserSwitchResult
 
         sut = PayPalWebLauncher("custom_url_scheme", liveConfig, browserSwitchClient)
-        val status = sut.completeAuthRequest(intent, "pending request")
-                as PayPalWebStatus.VaultSuccess
-        assertEquals("fake-approval-session-id", status.result.approvalSessionId)
+        val result = sut.completeVaultAuthRequest(intent, "pending request")
+                as PayPalWebCheckoutFinishVaultResult.Success
+        assertEquals("fake-approval-session-id", result.approvalSessionId)
     }
 
     @Test
-    fun `completeRequest() parses vault failure when approval session id is blank`() {
+    fun `completeVaultAuthRequest() parses vault failure when approval session id is blank`() {
         val browserSwitchResult = createVaultSuccessBrowserSwitchResult(
             requestCode = BrowserSwitchRequestCodes.PAYPAL_VAULT,
             setupTokenId = "fake-setup-token-id",
@@ -377,11 +340,11 @@ class PayPalWebLauncherUnitTest {
         } returns browserSwitchResult
 
         sut = PayPalWebLauncher("custom_url_scheme", liveConfig, browserSwitchClient)
-        val status = sut.completeAuthRequest(intent, "pending request")
-                as PayPalWebStatus.VaultError
+        val result = sut.completeVaultAuthRequest(intent, "pending request")
+                as PayPalWebCheckoutFinishVaultResult.Failure
         val expectedDescription =
             "Result did not contain the expected data. Payer ID or Order ID is null."
-        assertEquals(expectedDescription, status.error.errorDescription)
+        assertEquals(expectedDescription, result.error.errorDescription)
     }
 
     private fun createCheckoutMetadata(orderId: String) = JSONObject()